notsolved : redis 설정 분리 - 프로파일 (#12)
