+++ queue-server/src/main/java/com/j1p3ter/queueserver/config/ReactiveRedisConfig.java
@@ -1,17 +1,45 @@
 package com.j1p3ter.queueserver.config;
 
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.ReactiveRedisTemplate;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.RedisSerializationContext;
 import org.springframework.data.redis.serializer.RedisSerializer;
 
+
 @Configuration
 public class ReactiveRedisConfig {
+    @Value("${spring.data.redis.host}")
+    private String host;
+
+    @Value("${spring.data.redis.port}")
+    private int port;
+
+//    @Value("${spring.data.redis.username}")
+//    private String username;
+
+//    @Value("${spring.data.redis.password}")
+//    private String password;
+
+    @Bean(name = "reactiveRedisConnectionFactory")
+    public ReactiveRedisConnectionFactory reactiveRedisConnectionFactory() {
+        RedisStandaloneConfiguration configuration = new RedisStandaloneConfiguration();
+        configuration.setHostName(host);
+        configuration.setPort(port);
+//        configuration.setUsername(username);
+//        configuration.setPassword(password);
+        return new LettuceConnectionFactory(configuration);
+    }
+
     @Bean
-    public ReactiveRedisTemplate<String, Object> reactiveRedisTemplate(ReactiveRedisConnectionFactory factory) {
+    public ReactiveRedisTemplate<String,Object> reactiveRedisTemplate(
+            @Qualifier("reactiveRedisConnectionFactory") ReactiveRedisConnectionFactory factory) {
         Jackson2JsonRedisSerializer<Object> serializer = new Jackson2JsonRedisSerializer<>(Object.class);
         RedisSerializationContext.RedisSerializationContextBuilder<String, Object> builder =
                 RedisSerializationContext.newSerializationContext(RedisSerializer.string());
