--- queue-server/src/main/resources/application-dev.yaml
@@ -1,16 +1,0 @@
-spring:
-  config:
-    activate:
-      on-profile: dev
-  data:
-    redis:
-      host: redis
-      port: 6379
-      username: default
-      password: systempass
-
-eureka:
-  client:
-    service-url:
-      defaultZone: http://localhost:8761/eureka/
-
