+++ queue-server/src/main/resources/application-queue-dev.yaml
@@ -1,0 +1,20 @@
+spring:
+  config:
+    activate:
+      on-profile: queue-dev
+  data:
+    redis:
+      host: localhost
+      port: 6379
+#      username: default
+#      password: systempass
+
+eureka:
+  client:
+    register-with-eureka: true  # Eureka 서버에 등록
+    fetch-registry: true        # Eureka 서버로부터 서비스 조회
+    service-url:
+      defaultZone: http://localhost:8761/eureka/
+
+server:
+  host: localhost
