+++ queue-server/src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   profiles:
-    default: dev
+    default: queue-dev
   application:
     name: queue-server
